Implement rollup and uglify-es
Now I can write code across multiple files while rolling everything into one minified file for web!
