+++ index.html
@@ -53,7 +53,6 @@
     
     <script src="https://code.jquery.com/jquery-3.2.1.min.js" integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4=" crossorigin="anonymous"></script>
     <script src="//maxcdn.bootstrapcdn.com/bootstrap/3.3.1/js/bootstrap.min.js"></script>
-    <script src="webpage.js"></script>
-    <script src="src/japt.js"></script>
+    <script src="dist/webpage.min.js"></script>
   </body>
 </html>
